try kotlin

+++ src/main/java/cn/pgyyd/mcg/singleton/JobIDGenerator.java
@@ -16,7 +16,7 @@ public class JobIDGenerator {
     private JobIDGenerator() {
     }
 
-    public int generate() {
+    public long generate() {
         return atomicInteger.incrementAndGet();
     }
 }
